Reskin and Rejuvenate
It's been some three years since I updated this portfolio. In the light
of looking for where to keep my content (Medium is trying to enforce
payments to use the platform seamnessly, Wordpress is not geared towards
programming as much, other proprietary platforms want to own the post),
I graciously stumbled upon this repository and decided to revive it.

In this change, I

* updated the gems to remove the vulnerabilities
* reskinned the site with the Holo-Alfa Jekyll theme
* updated my working environment
* updated content in the about page
* added an archive page to list all the posts