updated _config.yml
In a bid to move this to netlify, I cchanged the URL to point to the
Netlify deployment. Ther are also some files which are excluded from the
deployment processing. This should hopefully speed up builds sometime in
the future. I don't know why I a catering for something far in the
future now, but hey... it's part of the theme, so why not?