Netlify site (#5)
* added netlify.toml

This is the base template for a Jekyll site hosted on Netlify. There is
more information on how to configure a Netlify site in their
[documentation](https://www.netlify.com/docs/netlify-toml-reference/)

* updated _config.yml

In a bid to move this to netlify, I cchanged the URL to point to the
Netlify deployment. Ther are also some files which are excluded from the
deployment processing. This should hopefully speed up builds sometime in
the future. I don't know why I a catering for something far in the
future now, but hey... it's part of the theme, so why not?

* add NetlifyCMS

Added the base configuration for the Netlify CMS. Let's see if this
thing is truly as great as people say it is.