Cleaning up my mac (#16)
* Add screenshots of the bloat

* add image for post cleanup

* Added the header image to the list of blog posts

I thought to add some images to the list of blog posts to make that page
more exciting. Currently, it's just a list of text, and that could get
boring really quick. Since the theme already had something for header
images, I thought to bring this into the list.

The markdown for this included abstracting the header file aspect into
its own `_include` file and bringing this into everwhere I needed it.
This way I wouldn't be duplicating code. That said though, I couldn't
find a good way to check if I am dealing with a `page` or a `post`
— quite honestly, I don't know why there is a disparity between these
two, but I have to deal with them. My ignorance of Jekyll forces me to
write bad Liquid code in the header-image includes file. But this works
for now; until there's a need (and the knowledge) to refactor this.

* Write up the blog post